+++ install-scripts/hyprwayland-scanner.sh
@@ -55,20 +55,23 @@ done
 
 printf "${NOTE} Installing hyprwayland-scanner...\n"  
 
-# Check if hyprwayland-scanner folder exists and remove it
-if [ -d "hyprwayland-scanner" ]; then
+# Check if hyprwayland-scanner folder exists and remove it (under build/src)
+SRC_DIR="$SRC_ROOT/hyprwayland-scanner"
+if [ -d "$SRC_DIR" ]; then
     printf "${NOTE} Removing existing hyprwayland-scanner folder...\n"
-    rm -rf "hyprwayland-scanner"
+    rm -rf "$SRC_DIR"
 fi
 
 # Clone and build hyprlang
 printf "${NOTE} Installing hyprwayland-scanner...\n"
-if git clone --recursive -b $tag https://github.com/hyprwm/hyprwayland-scanner.git; then
-    cd hyprwayland-scanner || exit 1
-	cmake -DCMAKE_INSTALL_PREFIX=/usr -B build
-	cmake --build build -j `nproc`
+if git clone --recursive -b $tag https://github.com/hyprwm/hyprwayland-scanner.git "$SRC_DIR"; then
+    cd "$SRC_DIR" || exit 1
+    BUILD_DIR="$BUILD_ROOT/hyprwayland-scanner"
+    mkdir -p "$BUILD_DIR"
+	cmake -DCMAKE_INSTALL_PREFIX=/usr -B "$BUILD_DIR"
+	cmake --build "$BUILD_DIR" -j `nproc`
     if [ $DO_INSTALL -eq 1 ]; then
-        if sudo cmake --install build 2>&1 | tee -a "$MLOG" ; then
+        if sudo cmake --install "$BUILD_DIR" 2>&1 | tee -a "$MLOG" ; then
             printf "${OK} hyprwayland-scanner installed successfully.\n" 2>&1 | tee -a "$MLOG"
         else
             echo -e "${ERROR} Installation failed for hyprwayland-scanner." 2>&1 | tee -a "$MLOG"
@@ -77,7 +80,7 @@ if git clone --recursive -b $tag https://github.com/hyprwm/hyprwayland-scanner.g
         echo "${NOTE} DRY RUN: Skipping installation of hyprwayland-scanner $tag."
     fi
     #moving the addional logs to Install-Logs directory
-    [ -f "$MLOG" ] && mv "$MLOG" ../Install-Logs/
+    [ -f "$MLOG" ] && mv "$MLOG" "$PARENT_DIR/Install-Logs/"
     cd ..
 else
     echo -e "${ERROR} Download failed for hyprwayland-scanner. Please check log." 2>&1 | tee -a "$LOG"